𝗟𝘂𝗰𝗸𝘆𝗡𝘂𝗺.𝗷𝗮𝘃𝗮
John has discovered a manuscript from ancient mathematicians that defines a "lucky number". An integer, k, is considered a lucky number if it can be expressed as the sum of distinct powers
of 7, i.e., k = a1 + a2 + ... + an, where each ai equals 7^p and p can be any positive integer. If i and j are distinct, then ai is not equal to aj.
For instance,
7 is a lucky number because it can be expressed as `7^1`.
56 is a lucky number because it can be expressed as the sum of `7^2` and `7^1`.
John has an array of n integers and he wants to determine how many elements in this array are lucky numbers. He needs your help to write a program that takes an integer n and an array 
consisting of n integers, and determines the quantity of lucky integers in this array.
Input:
The first line of input contains an integer n, which represents the number of elements in the array. The second line of input contains n space-separated integers.
Output:
The program should print the number of lucky integers in the given array.
Constraints:
1 <= n <= 100
1 <= arr[i] <= 1000000 (where arr[i] is the ith element of the array)
Example #1:
Input:
2
49 50
Output:
1
Explanation: 49 is a lucky number because it can be expressed as 7^2. However, 50 cannot be represented as a sum of distinct powers of 7.
Example #2:
Input:
2
7 49
Output:
2
Explanation: Both 7 and 49 are lucky numbers because they can be expressed as 7^1 and 7^2, respectively.